finish saving every frame while  parsing the dst video.

+++ mmaction/utils/misc.py
@@ -11,6 +11,7 @@ from typing import Optional, Union
 import cv2
 import mmcv
 import numpy as np
+from tqdm import tqdm
 
 
 def get_random_string(length: int = 15) -> str:
@@ -54,27 +55,31 @@ def frame_extract(video_path: str,
     frame_tmpl = osp.join(target_dir, 'img_{:06d}.jpg')
     assert osp.exists(video_path), f'file not exit {video_path}'
     vid = cv2.VideoCapture(video_path)
-    frames = []
+    frame_count = int(vid.get(cv2.CAP_PROP_FRAME_COUNT))
+    
+    # frames = []
     frame_paths = []
     flag, frame = vid.read()
     cnt = 0
     new_h, new_w = None, None
-    while flag:
-        if short_side is not None:
-            if new_h is None:
-                h, w, _ = frame.shape
-                new_w, new_h = mmcv.rescale_size((w, h), (short_side, np.Inf))
-            frame = mmcv.imresize(frame, (new_w, new_h))
+    with tqdm(total=frame_count) as pbar:
+        while flag:
+            if short_side is not None:
+                if new_h is None:
+                    h, w, _ = frame.shape
+                    new_w, new_h = mmcv.rescale_size((w, h), (short_side, np.Inf))
+                frame = mmcv.imresize(frame, (new_w, new_h))
 
-        frames.append(frame)
-        frame_path = frame_tmpl.format(cnt + 1)
-        frame_paths.append(frame_path)
+            # frames.append(frame)
+            frame_path = frame_tmpl.format(cnt + 1)
+            if os.path.exists(frame_path): continue
+            frame_paths.append(frame_path)
+            cv2.imwrite(frame_path, frame)
+            cnt += 1
+            flag, frame = vid.read()
+            pbar.update(1)
 
-        cv2.imwrite(frame_path, frame)
-        cnt += 1
-        flag, frame = vid.read()
-
-    return frame_paths, frames
+    return frame_paths
 
 
 class VideoWriter():
